Reduce code duplication in index.html.ep, and color output by row
Instead of setting background color for each column in the table
individually, set it for the whole row.  This allows to remove
some of code duplication: now setting columns that do not change
whether host is up or not is in one place.  The conditional got
split in two: separate for row color, separate for the contents
of the "Status" column.