Move checking if host is up (ping only for now) to Ajax
Instead of waiting in sequence for all pings and all connections to
succeed or fail before rendering the index page, put placeholders and
use 'service/ping/:id' from JavaScript (jQuery) with Ajax.

Note that for now the SSH connection check is dropped.  There is also
a bit of code duplication between Perl (the template) and JavaScript.