Move checking if host is up to JavaScript via XHR
Instead of waiting in sequence for all pings and all connections to
succeed or fail before rendering the index page, put placeholders and
use '/service/ping/:id' from JavaScript (jQuery) with Ajax.

Note that there is also a bit of code duplication between Perl code
(the template) and JavaScript code.