Allow to configure whether to use server-side or "client"-side checks
With the 'ajax' configuration option you can select to use Ajax method,
i.e. call '/service/ping/:id' from JavaScript, and modify the HTML.
The default is to use server-side check.

Note that with XHR, i.e. 'ajax' set, you might want to increase the
reload time, as generating the index is then near instantenous;
you won't see the '/wol/:id' page because it redirects immediately.

Note also that we need to use false-ish value different than undef
to denote that we checked the host, and it is not alive.